rename I24... module to i24...
